New functions
CPOptimality is the up-to-date version of  FisherInformationMatrix.R
EDOptimality is an empirical version of DOptimality.R
EKOptimality is an empirical version of KOptimality.R
The *_minimal functions are modifications of the glmssn functions in SSN
that are compatible with the design functions in this package.
stableInverse.R is a script for obtaining inverses for matrices with
small eigenvalues by eigenvalue decomposition.