Update to parallel functions
Now the parallel functions load SSN and SSNDesign (this package) into the parallel workers so that functions like glmssn_minimal can be called inside the parallel loop. This does seem to slow things down a fair bit... I will look into other solutions but for now this is good enough.