Major updates to deal with temporally replicated data
Previous versions of these functions assumed no temporal replicates. Now the function can deal with temporal replicates, though in the case of temporal replicates it may take much longer to evaluate U(d) for a design problem of the same size since most of these require the inversion of a covariance matrix.